+++ src/app/shared/shared.module.ts
@@ -6,6 +6,7 @@ import { FormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterModule } from '@angular/router';
+import { TranslateModule } from '@ngx-translate/core';
 import { NgxChartsModule } from '@swimlane/ngx-charts';
 import { MaterialModule } from './module/material/material.module';
 import { PoeModule } from './module/poe/poe.module';
@@ -23,6 +24,7 @@ import { PoeModule } from './module/poe/poe.module';
 
         // third party
         NgxChartsModule,
+        TranslateModule,
 
         // modules
         MaterialModule,